Move card flag rule parse/serialize to core module (#384)
The SettingsTab (framework layer) was importing from the task-agent
adapter, breaking the layering rule that framework must not depend on
a specific adapter. Move parseCardFlagRulesJson and serializeCardFlagRules
to core/cardFlags.ts where CardFlagRule is defined. The adapter module
becomes a thin re-export for backward compatibility.
